# openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Boring Things foundation",
    "version": "0.1.0",
    "description": "Owner-scoped manual Things. Import, discovery and assistant execution follow later."
  },
  "paths": {
...
    "/api/things:import": {
      "post": {
        "operationId": "startImport",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [],
        "responses": {
          "202": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportAccepted"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/InvalidInput"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportStart"
              }
            }
          }
        },
        "tags": ["Imports"],
        "summary": "Start import"
      }
    },
    "/api/imports/{id}": {
      "get": {
        "operationId": "getImport",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Id."
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Import"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/InvalidInput"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        },
        "tags": ["Imports"],
        "summary": "Get import"
      }
    },
    "/api/imports/{id}:confirm": {
      "post": {
        "operationId": "confirmImport",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Id."
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Import"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/InvalidInput"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportConfirmation"
              }
            }
          }
        },
        "tags": ["Imports"],
        "summary": "Confirm import"
      }
    },
    "/api/imports/{id}:retry": {
      "post": {
        "operationId": "retryImport",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Id."
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Import"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/InvalidInput"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        },
        "tags": ["Imports"],
        "summary": "Retry import"
      }
    },
    "/api/things/{thingId}/stream": {
      "get": {
        "operationId": "streamThing",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "thingId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Thing id."
          }
        ],
        "responses": {
          "200": {
            "description": "Full masked Thing snapshots; event: thing.snapshot; id: revision. Reconnect sends current persisted state.",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/InvalidInput"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        },
        "tags": ["Things"],
        "summary": "Stream thing"
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "message": {
            "type": "string"
          },
          "statusCode": {
            "type": "integer"
          }
        },
        "required": ["message", "statusCode"],
        "description": "Public failure message and HTTP status; private diagnostic details are omitted."
      },
...
      "ImportCandidate": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "categoryId": {
            "type": "string",
            "maxLength": 200,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]*$"
          }
        },
        "required": ["id", "name", "categoryId"],
        "description": "Detected Thing offered for owner confirmation."
      },
      "ImportUsage": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "model": {
            "type": "string"
          },
          "inputTokens": {
            "type": "integer",
            "minimum": 0
          },
          "outputTokens": {
            "type": "integer",
            "minimum": 0
          },
          "cachedTokens": {
            "type": "integer",
            "minimum": 0
          },
          "elapsedMs": {
            "type": "integer",
            "minimum": 0
          },
          "toolCalls": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": false,
              "properties": {
                "name": {
                  "type": "string"
                },
                "resultCount": {
                  "type": "integer"
                },
                "truncated": {
                  "type": "boolean"
                }
              },
              "required": ["name", "resultCount", "truncated"]
            }
          }
        },
        "required": [
          "model",
          "inputTokens",
          "outputTokens",
          "cachedTokens",
          "elapsedMs",
          "toolCalls"
        ],
        "description": "Provider usage and bounded tool execution totals for an attempt."
      },
      "Import": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "attachmentId": {
            "type": "string",
            "format": "uuid"
          },
          "thingId": {
            "type": ["string", "null"],
            "format": "uuid"
          },
          "status": {
            "$ref": "#/components/schemas/ImportStatusEnum"
          },
          "candidates": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ImportCandidate"
            }
          },
          "thingIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "error": {
            "type": ["string", "null"]
          },
          "usage": {
            "$ref": "#/components/schemas/ImportUsage"
          }
        },
        "required": [
          "id",
          "attachmentId",
          "thingId",
          "status",
          "candidates",
          "thingIds",
          "error",
          "usage"
        ],
        "description": "Public import state excluding raw extraction and private source content."
      },
      "ImportStart": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "attachmentId": {
            "type": "string",
            "format": "uuid"
          },
          "thingId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": ["attachmentId"],
        "description": "Source attachment and optional existing Thing to enrich."
      },
      "ImportAccepted": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "importId": {
            "type": "string",
            "format": "uuid"
          },
          "thingId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "$ref": "#/components/schemas/ImportStatusEnum"
          }
        },
        "required": ["importId", "thingId", "status"],
        "description": "Queued import identity and initial Thing identity."
      },
      "ImportConfirmation": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "selections": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": false,
              "properties": {
                "candidateId": {
                  "type": "string"
                },
                "targetThingId": {
                  "type": ["string", "null"],
                  "format": "uuid"
                }
              },
              "required": ["candidateId", "targetThingId"]
            },
            "minItems": 1,
            "maxItems": 10
          }
        },
        "required": ["selections"],
        "description": "Selected detected candidates and their optional existing targets."
      },
...
      "ImportStatusEnum": {
        "type": "string",
        "description": "Import status.",
        "enum": [
          "QUEUED",
          "EXTRACTING",
          "AWAITING_SELECTION",
          "MAPPING",
          "DISCOVERING",
          "COMPLETE",
          "INCOMPLETE",
          "FAILED"
        ]
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "A valid bearer token is required.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "NotFound": {
        "description": "The resource is missing, inaccessible or disabled.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Conflict": {
        "description": "The request conflicts with current state or an existing record.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "InvalidInput": {
        "description": "The request or a referenced value is invalid.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unavailable": {
        "description": "The requested capability is not configured or available.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "ServerError": {
        "description": "The request failed; private diagnostics are omitted.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    }
  },
...
}